ci: add github action that analyzes all python functions
This new github action runs a check on each the source branch
and the target branch and finds all functions that are missing
docstrings or functions that are missing type hints. And by
checking if the number of either of those two has increased in the
source branch, then it knows that new code was added that is
missing either a docstring or type hints.